% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_deg.r
\name{analyse_deg}
\alias{analyse_deg}
\title{Analyse Gene Expression Data (GEO Series)}
\usage{
analyse_deg(queryTerms, GSEaccession, GSEplatform)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}

\item{GSEaccession}{GSE Accession}

\item{GSEplatform}{GSE Platform}
}
\value{
A dataframe of differentially expressed protein-coding genes

A CSV-formmated file containing results of differentially expressed protein-coding genes (FilteredDE.csv)

A text file containing Entrez identifiers of differentially expressed protein-coding genes (DEGenes.txt)

A PDF-formatted file containing visualization plots (Plots.pdf)

A text file containing top 20 genes in tabulated form (Top20DEGenes.txt)
}
\description{
Perform differential expression analysis and retrieve differentially expressed protein-coding genes from GEO Series
}
\examples{
queryTerms <- c("endometriosis");
GSEaccession <- "GSE23339";
GSEplatform <- "GPL6102";
results <- analyse_deg(queryTerms,GSEaccession,GSEplatform);
}
