% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ppi_network.r
\name{analyse_ppi_network}
\alias{analyse_ppi_network}
\title{Analyse PPI Network}
\usage{
analyse_ppi_network(geneList)
}
\arguments{
\item{geneList}{A vector of Entrez gene identifiers}
}
\value{
A directory named PPI

PPI: A summary of PPI analysis (PPISummary.txt)

PPI: A PDF file containing visualization plots from PPI analysis (STRING_PPI_Network.pdf)

PPI: A list of filtered genes from PPI where isolated nodes are discarded (FilteredGenes.txt)

PPI: A list of gene clusters with details on enriched KEGG Pathway (GeneCluster.txt)

A directory named PPI/FilteredGenes containing functional enrichment results for filtered genes

A directory named PPI/TopGenes containing KEGG enrichment results for top 10 or 20 genes
}
\description{
A workflow to get predict protein-protein interaction, identify gene modules and perform functional enrichment for a given gene list
}
\examples{
gene_file <- 'DatabaseGenes.txt';
geneList <- readLines(gene_file)
results <- analyse_ppi_network(geneList);
}
