\name{check_libraries}
\alias{check_libraries}
\title{Check necessary libraries}
\description{Check, install and load necessary libraries}
\usage{check_libraries()}
\arguments{None}
\value{None}
\references{
https://www.disgenet.org/disgenet2r, https://github.com/ramiromagno/gwasrapidd, https://github.com/dlindholm/oboe/
}
\author{Nur Syatila Ab Ghani}
