% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialexpression_single_workflow.r
\name{differentialexpression_single_workflow}
\alias{differentialexpression_single_workflow}
\title{Single Differential Expression Worflow}
\usage{
differentialexpression_single_workflow(queryTerms, GSEaccession, GSEplatform)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords to be used for sample classification}

\item{GSEaccession}{GEO Series (GSE) accession}

\item{GSEplatform}{GEO Series (GSE) platform}
}
\value{
GeneExpression: A CSV-formmated file containing results of differentially expressed protein-coding genes (FilteredDE.csv) for individual GEO Series in a separate directory

GeneExpression: A text file containing Entrez identifiers of differentially expressed protein-coding genes (DEGenes.txt) for individual GEO Series in a separate directory

GeneExpression: A PDF-formatted file containing visualization plots (Plots.pdf) for individual GEO Series in a separate directory

GeneExpression: A text file containing top 20 genes in tabulated form (Top20DEGenes.txt) for individual GEO Series in a separate directory

PPI: A summary of PPI analysis (PPISummary.txt)

PPI: A list of filtered genes from PPI where isolated nodes are discarded (FilteredGenes.txt)

PPI: Top genes from PPI with highest interaction (Top_PPI_genes.txt)

PPI: A list of gene clusters with details on enriched KEGG Pathway (GeneCluster.txt)

PPI: A PDF file containing visualization plots from PPI analysis (STRING_PPI_Network.pdf)

Functional Enrichment: A separate directory of functional enrichment results for filtered genes (FilteredGenes/) and top genes (TopGenes) containing visualization plots (.pdf) and enrichment results (.csv)
}
\description{
Get differentially expressed genes from gene expression data associated to a disease, perform PPI and functional enrichment analyses for a single GEO Series
}
\examples{
queryTerms <- c("endometriosis");
GSEaccession <- "GSE23339";
GSEplatform <- "GPL6102";
results <- differentialexpression_single_workflow(queryTerms,GSEaccession,GSEplatform);
}
