% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesearch_workflow.r
\name{genesearch_workflow}
\alias{genesearch_workflow}
\title{Database Gene Search Worflow}
\usage{
genesearch_workflow(
  queryTerms,
  disgenet = FALSE,
  email = NULL,
  password = NULL
)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}

\item{disgenet}{An option to include results from DisGeNET. Note that This option will require email and password to DisGeNET account}

\item{email}{Registered email to DisGeNET account}

\item{password}{Password to DisGeNET account}
}
\value{
A directory named GeneSearch

Gene Search: CSV-formmated files containing results from individual database search

Gene Search: Text-formatted files containing a list of genes from individual database search ([Database]_Genes.txt)

Gene Search: Text-formatted files containing a list of protein-coding genes from individual database search ([Database]_PCGenes.txt)

Gene Search: A text-formatted file containing a list of protein-coding genes from all database searches (DatabaseGenes.txt)

PPI: A summary of PPI analysis (PPISummary.txt)

PPI: A list of filtered genes from PPI where isolated nodes are discarded (FilteredGenes.txt)

PPI: Top genes from PPI with highest interaction (Top_PPI_genes.txt)

PPI: A list of gene clusters with details on enriched KEGG Pathway (GeneCluster.txt)

PPI: A PDF file containing visualization plots from PPI analysis (STRING_PPI_Network.pdf)

Functional Enrichment: A separate directory of functional enrichment results
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from available public resources
}
\examples{
queryTerms <- c("endometriosis");
results <- genesearch_workflow(queryTerms,disgenet=FALSE);
results <- genesearch_workflow(queryTerms,disgenet=TRUE,"you@gmail.com","yourpassword");
}
