% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.r
\name{get_clinvar}
\alias{get_clinvar}
\title{Gene search in CLinVar}
\usage{
get_clinvar(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from ClinVar

A text-formatted file containing a list of genes (ClinVar_Genes.txt)

A text-formatted file containing a list of protein-coding genes (ClinVar_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from ClinVar (via rentrez)
}
\examples{
queryTerms <- c("endometriosis");
results <- get_clinvar(queryTerms);
}
