\name{get_disease_ontology_terms}
\alias{get_disease_ontology_terms}
\title{Retrieves Disease Ontology terms}
\description{Retrieves Disease Ontology terms related to a given query disease using oboe package}
\usage{get_disease_ontology_terms(queryTerms,summaryFile)}
\arguments{
\item{queryTerms}{Query terms in the form of list}
\item{summaryFile}{Summary file to record output}
}
\value{List of Disease Ontology terms and a summary file}
\references{
https://disease-ontology.org/
}
\author{Nur Syatila Ab Ghani}
