% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diseasesdb.r
\name{get_diseasesdb}
\alias{get_diseasesdb}
\title{Gene search in DISEASES database}
\usage{
get_diseasesdb(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from DISEASES database

A text-formatted file containing a list of genes (DISEASES_Genes.txt)

A text-formatted file containing a list of protein-coding genes (DISEASES_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from DISEASES database
}
\examples{
queryTerms <- c("endometriosis");
results <- get_diseasesdb(queryTerms);
}
