% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrichment_results.r
\name{get_enrichment_results}
\alias{get_enrichment_results}
\title{Get enrichment results}
\usage{
get_enrichment_results(geneList, ORAterm, enrichmenttype)
}
\arguments{
\item{geneList}{A vector / dataframe of Entrez gene identifiers}

\item{ORAterm}{Functional term, BP, MF, CC, KEGG or DO}

\item{enrichmenttype}{ORA (over-representation analysis) or GSE (gene set enrichment analysis)}
}
\value{
A CSV-formatted file of enrichment results

A PDF file containing visualization plots from enrichment analysis
}
\description{
Perform enrichment analysis for a gene list (over-representation analysis) or a dataframe consisting gene identifiers and log value (gene set enrichment analysis)
}
\examples{
#gene_file <- 'DatabaseGenes.txt';
#geneList <- readLines(gene_file)
geneList <- c("1588","3586","5241","7124", "998", "54361", "55591", "283455", "3480","8390","83608","9687","4211");
results <- get_enrichment_results(geneList,"KEGG","ORA");
}
