% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_modules.r
\name{get_gene_modules}
\alias{get_gene_modules}
\title{Get gene modules}
\usage{
get_gene_modules(geneNetwork)
}
\arguments{
\item{geneNetwork}{An igraph object}

\item{summaryFile}{A file name to store details of the clusters}
}
\value{
PPI: A list of gene clusters with details on enriched KEGG Pathway (summaryFile)
}
\description{
Get gene modules (clusters of genes according to interactions) and analyse individual gene clusters/modules (required by analyse_ppi_network())
}
\examples{
geneList <- readLines(gene_file)
genelist <- as.data.frame(geneList)
mapped_genes <- suppressWarnings(string_db$map(genelist, "geneList", removeUnmappedRows = TRUE, quiet = TRUE ))
string_db <- STRINGdb$new( version="11.5", species=9606, score_threshold=700, input_directory=".")
gene_network <- suppressMessages(string_db$get_subnetwork(mapped_genes$STRING_id))
gene_network2 <- igraph::delete.vertices(gene_network, degree(gene_network)==0)
gene_network2 <- igraph::simplify(gene_network2, remove.multiple=TRUE, remove.loops=TRUE)
results <- get_gene_modules(gene_network2)
}
