% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gse_datasets.r
\name{get_gse_datasets}
\alias{get_gse_datasets}
\title{Get GSE datasets associated to a disease}
\usage{
get_gse_datasets(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A dataframe of results (GEO Dataset with 'diseased state' condition) from GEO Dataset web search through Rentrez

A CSV-formmated file containing GEO Datasets associated to a disease (GSESummary.csv)

A CSV-formmated file containing GEO Datasets associated to a disease with 'diseased state' condition (DiseaseGSESummary.csv)
}
\description{
Retrieve GSE Datasets associated to a disease based on MeSH Terms
}
\examples{
queryTerms <- c("endometriosis");
results <- get_gse_datasets(queryTerms);
}
