% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtr.r
\name{get_gtr}
\alias{get_gtr}
\title{Gene search in GTR}
\usage{
get_gtr(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from GTR database

A text-formatted file containing a list of genes (GTR_Genes.txt)

A text-formatted file containing a list of protein-coding genes (GTR_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from GTR (via rentrez)
}
\examples{
queryTerms <- c("endometriosis");
results <- get_gtr(queryTerms);
}
