% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gwascatalog.r
\name{get_gwascatalog}
\alias{get_gwascatalog}
\title{Gene search in GWAS Catalog}
\usage{
get_gwascatalog(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from GWAS Catalog database

A text-formatted file containing a list of genes (GWASCatalog_Genes.txt)

A text-formatted file containing a list of protein-coding genes (GWASCatalog_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from GWAS Catalog (via gwasrapidd)
}
\examples{
queryTerms <- c("endometriosis");
results <- get_gwascatalog(queryTerms);
}
