% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_medgen.r
\name{get_medgen}
\alias{get_medgen}
\title{Gene search in MedGen}
\usage{
get_medgen(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from MedGen

A text-formatted file containing a list of genes (MedGen_Genes.txt)

A text-formatted file containing a list of protein-coding genes (MedGen_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from MedGen (via Rentrez)
}
\examples{
queryTerms <- c("endometriosis");
results <- get_medgen(queryTerms);
}
