% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meshterms.r
\name{get_meshterms}
\alias{get_meshterms}
\title{Get MeSH terms or identifiers}
\usage{
get_meshterms(queryTerms, meshtype)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}

\item{meshtype}{type of MeSH output, "mesh_id" or "mesh_term"}
}
\value{
A list of MeSH terms or identifers
}
\description{
Get MeSH terms or identifiers associated to a disease (via rentrez)
}
\examples{
queryTerms <- c("endometriosis");
results <- get_meshterms(queryTerms,"mesh_id"),or;
results <- get_meshterms(queryTerms,"mesh_term");
}
