% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omim.r
\name{get_omim}
\alias{get_omim}
\title{Gene search in OMIM}
\usage{
get_omim(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from OMIM

A text-formatted file containing a list of genes (OMIM_Genes.txt)

A text-formatted file containing a list of protein-coding genes (OMIM_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from OMIM (via rentrez)
}
\examples{
queryTerms <- c("endometriosis");
results <- get_omim(queryTerms);
}
