% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PCGenes.r
\name{get_PCGenes}
\alias{get_PCGenes}
\title{Get protein coding genes}
\usage{
get_PCGenes(geneList, gene_file, pcgene_file, key_type)
}
\arguments{
\item{geneList}{A vector of genes (identifiers/symbols)}

\item{gene_file}{A text-formatted file name to store the list of genes in tabular form (ENTREZID, SYMBOL, GENETYPE)}

\item{pcgene_file}{A text-formatted file name to store the list of protein-coding genes (Entrez identifiers)}

\item{key_type}{A text-formatted file name to store the list of protein-coding genes (Entrez identifiers)}
}
\value{
Type of geneList, e.g. "ENSEMBL","ENTREZID","SYMBOL","UNIPROT" etc. Refer to keytypes(org.Hs.eg.db)
}
\description{
Get protein coding genes from a list of genes
}
\examples{
geneList <- c("1588","3586","5241","100048912","10151");
gene_file <- "Genes.txt"
pcgene_file <- "PCGenes.txt"
results <- get_PCGenes(geneList,gene_file,pcgene_file,"ENTREZID");
}
