% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phegeni.r
\name{get_phegeni}
\alias{get_phegeni}
\title{Gene search in PheGenI}
\usage{
get_phegeni(queryTerms)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}
}
\value{
A list of protein-coding genes

A CSV-formmated file containing results from PheGenI

A text-formatted file containing a list of genes (PheGenI_Genes.txt)

A text-formatted file containing a list of protein-coding genes (PheGenI_PCGenes.txt)
}
\description{
Get disease-associated genes (Entrez Gene identifiers) from PheGenI
}
\examples{
queryTerms <- c("endometriosis");
results <- get_phegeni(queryTerms);
}
