\name{get_ppi_plot}
\alias{get_ppi_plot}
\title{Plot a network using igraph package}
\description{Plot a network based on layout.fruchterman.reingold layout.}
\usage{get_ppi_plot(geneNetwork,graphTitle,geneAnnotation)}
\arguments{
\item{geneNetwork}{Gene network for plotting}
\item{graphTitle}{Graph title}
\item{geneAnnotation}{Gene annotation}
}
\value{Network plot}
\references{
https://cran.r-project.org/web/packages/igraph/index.html
}
\author{Nur Syatila Ab Ghani}
