% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samplegroupings.r
\name{get_samplegroupings}
\alias{get_samplegroupings}
\title{Assign sample groupings for a GEO Series)}
\usage{
get_samplegroupings(queryTerms, GSEaccession, GSEplatform)
}
\arguments{
\item{queryTerms}{A vector of disease name and/or keywords}

\item{GSEaccession}{GSE Accession}

\item{GSEplatform}{GSE Platform}
}
\value{
A vector of GEOquery results and sample groupings e.g. c(0,0,0,0,1,1,1,1)
}
\description{
Assign sample groupings for a GEO Series through automatic classification (required by analyse_deg function)
}
\examples{
queryTerms <- c("endometriosis");
GSEaccession <- "GSE23339";
GSEplatform <- "GPL6102";
results <- get_samplegroupings(queryTerms,GSEaccession,GSEplatform);
}
